updated product service

+++ product-service/src/main/java/com/aithmetic/product/dto/ProductResponse.java
@@ -18,5 +18,5 @@ public class ProductResponse {
     private String name;
     private String description;
     private BigDecimal price;
-    private BigInteger quantity;
+    private int quantity;
 }
