+++ product-service/src/main/java/com/aithmetic/product/model/Product.java
@@ -27,5 +27,5 @@ public class Product {
     private String name;
     private String description;
     private BigDecimal price;
-    private BigInteger quantity;
+    private int quantity;
 }
